+++ CHANGELOG.md
@@ -54,6 +54,7 @@
   ([GH-1211](https://github.com/NVIDIA/warp/issues/1211)).
 - Fix `@wp.func` losing parameter type information in Pyright/Pylance
   ([GH-1219](https://github.com/NVIDIA/warp/issues/1219)).
+- Fix gradients not propagating through array-of-structs ([GH-1207](https://github.com/NVIDIA/warp/pull/1207)).
 - Fix crashes caused by C++ standard library mismatch on Linux.
 
 ### Documentation